move constants to separate file, away from helpers
