+++ Assessments.Frontend.Web/Views/AlienSpecies/2023/AssessmentPartials/_TableOfContents.cshtml
@@ -15,7 +15,7 @@
             foreach (var heading in Constants.HeadingsNO)
             {
                 
-               <!--<li is-visible=@Constants.showHeading(@heading.Key,@Model.Assessment.Category)>-->
+               <!--<li is-visible=@Constants.ShowHeading(@heading.Key,@Model.Assessment.Category)>-->
                 <li>
                     <a href="#@heading.Key">@heading.Value</a>
